+++ game/tile.h
@@ -3,7 +3,7 @@
 
 extern unsigned short* tile_tilePtr;
 extern unsigned short* tile_itemPtr;
-extern int tile_tileX;
+extern int16_t tile_tileX;
 
 #define TILE_SKY 0xbba6
 #define TILE_COLLISION(x) (x < 0x1900)
@@ -15,7 +15,7 @@ tile_init(void);
 void 
 tile_renderScreen(void);
 void
-tile_invalidateTile(int x, int y, int offset);
+tile_invalidateTile(int16_t x, int16_t y, int16_t offset);
 void
 tile_renderInvalidTiles(frame_buffer_t fb);
 
@@ -32,10 +32,10 @@ tile_renderNextTileDown(uint16_t hscroll);
 typedef struct tile_redraw {
   struct tile_redraw* prev;
   struct tile_redraw* next;
-  int x;
-  int y;
-  int offset;
-  int count;
+  int16_t x;
+  int16_t y;
+  int16_t offset;
+  int16_t count;
 } tile_redraw_t;
